requirements.txt
Django>=2.2.8,<3
djangorestframework>=3.10.0,<=3.11.0
flake8==3.7.9
requests==2.23.0
mysqlclient